requirements.txt
docker
docker-compose
flask
flask_executor
netaddr
pyyaml
uwsgidecorators